Refactored files to put all models into the models folder, moved from providers.
