+++ frontend_scv/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:unison/screens/MessageTest.dart';
 import 'package:unison/screens/judge_duty_screen.dart';
 import 'package:unison/screens/messaging_screen.dart';
 
@@ -9,7 +10,7 @@ import './screens/splash_screen.dart';
 import './screens/edit_contract_screen.dart';
 import './screens/view_contract_screen.dart';
 //
-import './providers/contracts.dart';
+import 'models/contracts.dart';
 import './providers/auth.dart';
 
 void main() => runApp(MyApp());
