+++ frontend_scv/lib/models/condition.dart
