+++ frontend_scv/lib/models/contracts.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
-import 'package:unison/providers/global.dart';
+import 'package:unison/models/global.dart';
 import 'package:unison/services/Server/commonService.dart';
-import './contract.dart';
-import '../models/http_exception.dart';
+import 'contract.dart';
+import 'http_exception.dart';
 
 class Contracts with ChangeNotifier {
   List<Contract> _items = [];
