+++ frontend_scv/lib/models/global.dart
