+++ frontend_scv/lib/models/message.dart
