+++ frontend_scv/lib/screens/view_contract_screen.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:unison/providers/condition.dart';
-import 'package:unison/providers/global.dart';
+import 'package:unison/models/condition.dart';
+import 'package:unison/models/global.dart';
 import 'package:unison/screens/messaging_screen.dart';
 import 'package:unison/services/Server/negotiationService.dart';
 import '../widgets/contract_conditions_panel.dart';
 import '../widgets/contract_detail_info_panel.dart';
-import '../providers/contracts.dart';
+import '../models/contracts.dart';
 
 class ViewContractScreen extends StatefulWidget {
   static const routeName = '/view-contract';
