+++ frontend_scv/lib/widgets/contract_conditions_panel.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:unison/services/Server/negotiationService.dart';
 import 'package:unison/widgets/condition_item.dart';
-import '../providers/contract.dart';
+import '../models/contract.dart';
 
 class ContractConditionsPanel extends StatelessWidget {
   final Contract _contract;
