+++ frontend_scv/lib/widgets/contract_item.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import '../providers/auth.dart';
-import '../providers/contract.dart';
+import '../models/contract.dart';
 import '../screens/view_contract_screen.dart';
 
 class ContractItem extends StatelessWidget {
