+++ frontend_scv/lib/widgets/judge_active_cases_grid.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:unison/widgets/judge_active_contract_item.dart';
-import '../providers/contracts.dart';
+import '../models/contracts.dart';
 
 class JudgeActiveCasesGrid extends StatelessWidget {
   JudgeActiveCasesGrid();
