+++ frontend_scv/lib/widgets/message_item.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:unison/providers/condition.dart';
-import 'package:unison/providers/contracts.dart';
-import 'package:unison/providers/global.dart';
-import 'package:unison/providers/message.dart';
+import 'package:unison/models/condition.dart';
+import 'package:unison/models/contracts.dart';
+import 'package:unison/models/global.dart';
+import 'package:unison/models/message.dart';
 import 'package:unison/services/Server/messageService.dart';
 import 'package:unison/services/Server/negotiationService.dart';
